+++ app/admin/categories.rb
@@ -6,15 +6,15 @@ ActiveAdmin.register Category do
   #
   # Uncomment all parameters which should be permitted for assignment
   #
-  permit_params :organization_id, :name
+  # permit_params :organization_id, :name
   #
   # or
   #
-  # permit_params do
-  #   permitted = [:organization_id, :name]
-  #   permitted << :other if params[:action] == 'create' && current_user.admin?
-  #   permitted
-  # end
+  permit_params do
+    permitted = [:name]
+    permitted << :organization_id if current_admin_user.is_root?
+    permitted
+  end
 
   controller do
     def scoped_collection
@@ -23,4 +23,21 @@ ActiveAdmin.register Category do
       myscope
     end
   end
+
+  index do
+    selectable_column
+    id_column
+    if current_admin_user.is_root?
+      column(:organization)
+    end
+    column(:name)
+    column(:created_at)
+    column(:updated_at)
+    actions
+  end
+
+  filter :organization, if: proc { current_admin_user.is_root? }
+  filter :name
+  filter :created_at
+  filter :updated_at
 end