+++ app/admin/trainers.rb
@@ -11,7 +11,7 @@ ActiveAdmin.register Trainer do
 
   permit_params do
     permitted = [:firstname, :lastname, :nickname, :bio, :state]
-    # permitted << :other if params[:action] == 'create' && current_user.admin?
+    permitted << :organization_id if current_admin_user.is_root?
     permitted
   end
 
@@ -22,4 +22,29 @@ ActiveAdmin.register Trainer do
       myscope
     end
   end
+
+  index do
+    selectable_column
+    id_column
+    if current_admin_user.is_root?
+      column(:organization)
+    end
+    column(:firstname)
+    column(:lastname)
+    column(:nickname)
+    column(:bio)
+    column(:state) {|obj| span obj.state, class: "status_tag #{obj.state}" }
+    column(:created_at)
+    column(:updated_at)
+    actions
+  end
+
+  filter :organization    , if: proc { current_admin_user.is_root? }
+  filter :firstname
+  filter :lastname
+  filter :nickname
+  filter :bio
+  filter :state       , as: :select, collection: Trainer.localized_states
+  filter :created_at
+  filter :updated_at
 end