+++ app/models/organization.rb
@@ -1,8 +1,13 @@
 class Organization < ApplicationRecord
+  include Stateable
+
   has_many :users, dependent: :destroy
   has_many :courses, dependent: :destroy
   has_many :trainers, dependent: :destroy
   has_many :rooms, dependent: :destroy
+  has_many :categories, dependent: :destroy
+  has_many :products, dependent: :destroy
+  has_many :subscriptions, dependent: :destroy
   has_many :system_logs, dependent: :destroy
 
   STATES = { activating: 10, active: 20, suspended: 30}