+++ db/migrate/20200111155410_devise_create_admin_users.rb
@@ -3,6 +3,8 @@
 class DeviseCreateAdminUsers < ActiveRecord::Migration[6.0]
   def change
     create_table :admin_users do |t|
+      t.references :organization, null: true, foreign_key: true
+
       ## Database authenticatable
       t.string :email,              null: false, default: ""
       t.string :encrypted_password, null: false, default: ""
@@ -31,7 +33,7 @@ class DeviseCreateAdminUsers < ActiveRecord::Migration[6.0]
       # t.integer  :failed_attempts, default: 0, null: false # Only if lock strategy is :failed_attempts
       # t.string   :unlock_token # Only if unlock strategy is :email or :both
       # t.datetime :locked_at
-
+      t.integer "roles_mask"
 
       t.timestamps null: false
     end